Renamed some files for uniformity;
