+++ Source/Open World - Server/Open World Server/Server_Vars.cs
